+++ packages/modules/product/src/services/product-module-service.ts
@@ -59,8 +59,8 @@ import {
   UpdateTypeInput,
   VariantImageInputArray,
 } from "../types"
-import { joinerConfig } from "./../joiner-config"
 import { eventBuilders } from "../utils/events"
+import { joinerConfig } from "./../joiner-config"
 
 type InjectedDependencies = {
   baseRepository: DAL.RepositoryService
@@ -1744,12 +1744,15 @@ export default class ProductModuleService
         .registerSubscriber(new subscriber(sharedContext))
     }
 
+    const productIds = data.map((d) => d.id).filter(Boolean)
+
     const originalProducts = await this.productService_.list(
       {
-        id: data.map((d) => d.id),
+        id: productIds,
       },
       {
-        relations: ["options", "options.values", "variants", "images", "tags"],
+        relations: ["options", "options.values", "tags"],
+        take: productIds.length,
       },
       sharedContext
     )
@@ -2013,10 +2016,9 @@ export default class ProductModuleService
       // Re map options to handle non serialized data as well
       dbOptions =
         originalProducts
-          ?.map((originalProduct) =>
+          ?.flatMap((originalProduct) =>
             originalProduct.options.map((option) => option)
           )
-          .flat()
           .filter(Boolean) ?? []
     }